Fixed edit functionality for Clients, Services, and Projects.

+++ FreelanceTimeTracker/Controllers/ProjectsController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using FreelanceTimeTracker.Models;
+using Microsoft.AspNet.Identity;
 
 namespace FreelanceTimeTracker.Controllers
 {
@@ -15,20 +16,29 @@ namespace FreelanceTimeTracker.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Projects
+        [Authorize]
         public ActionResult Index()
         {
-            return View(db.Projects.ToList());
+            var userName = User.Identity.GetUserName();
+
+            var projects = from p in db.Projects select p;
+            var thisUsersProjects = projects.Where(p => p.Client.ClientOwner.Equals(userName));
+
+            return View(thisUsersProjects);
         }
 
         // GET: Projects/Details/5
+        [Authorize]
         public ActionResult Details(int? id)
         {
+            var userName = User.Identity.GetUserName();
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Project project = db.Projects.Find(id);
-            if (project == null)
+            if (project == null || !project.Client.ClientOwner.Equals(userName))
             {
                 return HttpNotFound();
             }
@@ -36,9 +46,16 @@ namespace FreelanceTimeTracker.Controllers
         }
 
         // GET: Projects/Create
+        [Authorize]
         public ActionResult Create()
         {
-            return View();
+            var usersClients = GetUsersClients();
+
+            var projectModel = new Project();
+
+            projectModel.Clients = GetSelectedListItems(usersClients);
+
+            return View(projectModel);
         }
 
         // POST: Projects/Create
@@ -46,8 +63,12 @@ namespace FreelanceTimeTracker.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ProjectID,ProjectName")] Project project)
+        [Authorize]
+        public ActionResult Create([Bind(Include = "ProjectID,ProjectName,SelectedClient")] Project project)
         {
+
+            project.ClientID = Convert.ToInt32(project.SelectedClient);
+
             if (ModelState.IsValid)
             {
                 db.Projects.Add(project);
@@ -59,14 +80,17 @@ namespace FreelanceTimeTracker.Controllers
         }
 
         // GET: Projects/Edit/5
+        [Authorize]
         public ActionResult Edit(int? id)
         {
+            var userName = User.Identity.GetUserName();
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Project project = db.Projects.Find(id);
-            if (project == null)
+            if (project == null || !project.Client.ClientOwner.Equals(userName))
             {
                 return HttpNotFound();
             }
@@ -78,11 +102,22 @@ namespace FreelanceTimeTracker.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ProjectID,ProjectName")] Project project)
+        [Authorize]
+        public ActionResult Edit([Bind(Include = "ProjectID,ProjectName,ClientID")] Project project)
         {
+            var userName = User.Identity.GetUserName();
+
+            var dbClient = db.Clients.Where(c => c.ClientID == project.ClientID).ToList()[0]; // There can only be one client with this ID.
+
+            if (!dbClient.ClientOwner.Equals(userName))
+            {
+                return View(project);
+            }
+            project.Client = dbClient;
             if (ModelState.IsValid)
             {
                 db.Entry(project).State = EntityState.Modified;
+                //db.Projects.Attach(project);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
@@ -90,14 +125,17 @@ namespace FreelanceTimeTracker.Controllers
         }
 
         // GET: Projects/Delete/5
+        [Authorize]
         public ActionResult Delete(int? id)
         {
+            var userName = User.Identity.GetUserName();
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Project project = db.Projects.Find(id);
-            if (project == null)
+            if (project == null || !project.Client.ClientOwner.Equals(userName))
             {
                 return HttpNotFound();
             }
@@ -107,11 +145,18 @@ namespace FreelanceTimeTracker.Controllers
         // POST: Projects/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult DeleteConfirmed(int id)
         {
+            var userName = User.Identity.GetUserName();
+
             Project project = db.Projects.Find(id);
-            db.Projects.Remove(project);
-            db.SaveChanges();
+            if (project.Client.ClientOwner.Equals(userName))
+            {
+                db.Projects.Remove(project);
+                db.SaveChanges();
+            }
+            
             return RedirectToAction("Index");
         }
 
@@ -122,6 +167,31 @@ namespace FreelanceTimeTracker.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        private ICollection<Client> GetUsersClients()
+        {
+            var userName = User.Identity.GetUserName();
+
+            var clientList = db.Clients.Where(c => c.ClientOwner.Equals(userName));
+
+            return clientList.ToList();
+        }
+
+        private IEnumerable<SelectListItem> GetSelectedListItems(IEnumerable<Client> elements)
+        {
+            var selectedList = new List<SelectListItem>();
+
+            foreach(var element in elements)
+            {
+                selectedList.Add(new SelectListItem
+                {
+                    Value = element.ClientID.ToString(),
+                    Text = element.ClientName
+                });
+            }
+
+            return selectedList;
         }
     }
 }
